Inline formatDateTimeForDisplay into DateTime.ts
Matches the convention used by Url.ts (formatUrlForDisplay) and
Select.ts (resolveSelectLabel), where presentation helpers live
alongside the property type rather than in a sibling file.
